Update Documentation, Add Stack Management
  * BUILD.md
    - Add build documentation for tested/confirmed systems

  * README.md
    - Fix typo in pronunciation
    - Reference BUILD.md

  * examples/esrc.newnix
    - Create generic push/pop stack management functions
    - Use push/pop functions to implement `pushd`/`popd`/`dirs`

  * prim-etc.c
    - Change `sum` and `sub` functions to use `ssize_t` instead of
      `int64_t` to test unexpected integer values when running code like
      `{sub 1}`

+++ prim-etc.c
@@ -32,9 +32,9 @@ PRIM(echo) {
  * after tackling the tail call recursion issue mentioned in "TODO"
  */
 PRIM(sum) {
-	int64_t sum = 0;
+	ssize_t sum = 0;
 	for (; list != NULL; list = list->next) {
-		sum += (int64_t)strtol(getstr(list->term), (char **)NULL, 10);
+		sum += (ssize_t)strtol(getstr(list->term), (char **)NULL, 10);
 	}
 	return mklist(mkstr(str("%ld", sum)), NULL);
 }
@@ -45,12 +45,12 @@ PRIM(sum) {
  * `0 - -1` with a resulting value of 1.
  */
 PRIM(sub) {
-	int64_t sum = 0;
+	ssize_t sum = 0;
 
 	if (list != NULL) {
-		sum = (int64_t)strtol(getstr(list->term), (char **)NULL, 10);
+		sum = (ssize_t)strtol(getstr(list->term), (char **)NULL, 10);
 		for (list = list->next; list != NULL; list = list->next) {
-			sum -= (int64_t)strtol(getstr(list->term), (char **)NULL, 10);
+			sum -= (ssize_t)strtol(getstr(list->term), (char **)NULL, 10);
 		}
 	}
 	return mklist(mkstr(str("%ld", sum)), NULL);
